Add NetworkManager & Divided local and network function
